**What is this feature?**
Adds support for installing 2 or more levels of plugin dependencies. For example the `redis-explorer-app` has a dependency on the `redis-app`. This is supported already today. However, `redis-app` has a dependency on `redis-datasource` (IE `redis-explorer-app` -> `redis-app` -> `redis-datasource`). This pull request ensures that we also install this plugins (and beyond if necessary).

Before:
```
INFO [07-08|11:06:38] Downloaded and extracted redis-explorer-app v2.1.1 zip successfully to /Users/wbrowne/dev/grafana/data/plugins/redis-explorer-app logger=installer.fs
INFO [07-08|11:06:38] Fetching redis-app dependencies...       logger=plugin.installer
INFO [07-08|11:06:38] Downloaded and extracted redis-app v2.2.1 zip successfully to /Users/wbrowne/dev/grafana/data/plugins/redis-app logger=installer.fs
DEBUG[07-08|11:06:38] Loading plugin source...                 logger=plugin.loader source=external
INFO [07-08|11:06:38] Plugin registered                        logger=plugins.registration pluginId=redis-explorer-app
INFO [07-08|11:06:38] Plugin registered                        logger=plugins.registration pluginId=redis-enterprise-software-datasource
INFO [07-08|11:06:38] Plugin registered                        logger=plugins.registration pluginId=redis-app
INFO [07-08|11:06:38] Plugin registered                        logger=plugins.registration pluginId=redis-latency-panel
INFO [07-08|11:06:38] Plugin registered                        logger=plugins.registration pluginId=redis-cli-panel
INFO [07-08|11:06:38] Plugin registered                        logger=plugins.registration pluginId=redis-cpu-panel
INFO [07-08|11:06:38] Plugin registered                        logger=plugins.registration pluginId=redis-gears-panel
INFO [07-08|11:06:38] Plugin registered                        logger=plugins.registration pluginId=redis-keys-panel
DEBUG[07-08|11:06:38] Plugin source loaded                     logger=plugin.loader source=external plugins="redis-app, redis-cli-panel, redis-cpu-panel, redis-enterprise-software-datasource, redis-explorer-app, redis-gears-panel, redis-keys-panel, redis-latency-panel" duration=16.727792ms
```

After:
```
INFO [07-08|11:04:25] Downloaded and extracted redis-explorer-app v2.1.1 zip successfully to /Users/wbrowne/dev/grafana/data/plugins/redis-explorer-app logger=installer.fs
INFO [07-08|11:04:25] Fetching redis-explorer-app dependency redis-app... logger=plugin.installer
INFO [07-08|11:04:26] Downloaded and extracted redis-app v2.2.1 zip successfully to /Users/wbrowne/dev/grafana/data/plugins/redis-app logger=installer.fs
INFO [07-08|11:04:26] Fetching redis-app dependency redis-datasource... logger=plugin.installer
INFO [07-08|11:04:31] Downloaded and extracted redis-datasource v2.1.1 zip successfully to /Users/wbrowne/dev/grafana/data/plugins/redis-datasource logger=installer.fs
DEBUG[07-08|11:04:31] Loading plugin source...                 logger=plugin.loader source=external
INFO [07-08|11:04:31] Plugin registered                        logger=plugins.registration pluginId=redis-datasource
DEBUG[07-08|11:04:31] Plugin source loaded                     logger=plugin.loader source=external plugins=redis-datasource duration=224.518125ms
DEBUG[07-08|11:04:31] Loading plugin source...                 logger=plugin.loader source=external
INFO [07-08|11:04:32] Plugin registered                        logger=plugins.registration pluginId=redis-app
INFO [07-08|11:04:32] Plugin registered                        logger=plugins.registration pluginId=redis-keys-panel
INFO [07-08|11:04:32] Plugin registered                        logger=plugins.registration pluginId=redis-latency-panel
INFO [07-08|11:04:32] Plugin registered                        logger=plugins.registration pluginId=redis-cli-panel
INFO [07-08|11:04:32] Plugin registered                        logger=plugins.registration pluginId=redis-cpu-panel
INFO [07-08|11:04:32] Plugin registered                        logger=plugins.registration pluginId=redis-gears-panel
DEBUG[07-08|11:04:32] Plugin source loaded                     logger=plugin.loader source=external plugins="redis-app, redis-cli-panel, redis-cpu-panel, redis-gears-panel, redis-keys-panel, redis-latency-panel" duration=13.8365ms
DEBUG[07-08|11:04:32] Loading plugin source...                 logger=plugin.loader source=external
INFO [07-08|11:04:32] Plugin registered                        logger=plugins.registration pluginId=redis-explorer-app
INFO [07-08|11:04:32] Plugin registered                        logger=plugins.registration pluginId=redis-enterprise-software-datasource
DEBUG[07-08|11:04:32] Plugin source loaded                     logger=plugin.loader source=external plugins="redis-enterprise-software-datasource, redis-explorer-app" duration=8.367375ms
```

**Why do we need this feature?**

While there is only one plugin (`redis-explorer-app`) that relies on this functionality - this should be supported for correctness.

**Who is this feature for?**

Plugin users / operators.

<!--

**Which issue(s) does this PR fix?**:


- Automatically closes linked issue when the Pull Request is merged.

Usage: "Fixes #<issue number>", or "Fixes (paste link of issue)"



Fixes #

-->
**Special notes for your reviewer:**
These changes are reflected in both the grafana CLI and the plugin catalog.