@wbrowne 2024-07-08
> LGTM. Is there a chance for circular dependencies that we need to cater for? 🤔

Yes 😉 that exists today. Will create an issue for it!