@marefr 2024-07-08
> > LGTM. Is there a chance for circular dependencies that we need to cater for? 🤔
> 
> Yes 😉 ~that exists today~. Will create an issue for it and do it in a follow up.

But today it's maximum 1 level right so can't see that happen? If not a big chance happening might be worth punt until someone experience it :)